Add a workaround for semaphore failing when system time changes
If the system time changes during the call to vcos_semaphore_wait, then the call will fail with EAGAIN even though the wait has not been running for the full length of the timeout. See:

https://github.com/pikvm/ustreamer/issues/56#issuecomment-726587254

As a workaround, we're temporarily ignoring the return value on the semaphore wait so that we don't prematurely abandon OMX encoding on one semaphore wait failure.